Parameters used in multiple functions

Reused parameter definitions are gathered here for easier usage.

Arguments

colData

character The name of a column that contains numeric data.

colGroup

character The name of a column to group by.

colPlace

character The name of a column with integer values between 0 and 9.

dfData

data.frame A data.frame with a grouping column and at least one numeric data column.

dfDigits

data.frame A data.frame with a grouping column and at least one column for a digit place.

max_power_10

integer The highest 10s place to check, as a power of 10.

min_power_10

integer The lowest 10s place to check, as a power of 10. Must be lower than max_power_10.

power_10

integer The 10s place to check, as a power of 10.

scaleDigitPalette

An optional ggplot2::scale_fill_discrete() palette to use for the digits.

strGroupLabel

character An optional name to use for the grouping column in labels.

strPlaceLabel

character An optional name to use for the place column in labels.

strChartTitle

character An optional title for the chart.

strCase

character Whether to display text in "Title" case, "lower" case, or "UPPER" case.

themePlot

An optional ggplot2::theme() to apply to the plot.

x

numeric A vector of values to evaluate.

x_arg

character The name of x to use in the output table and error messages. By default this is automatically determined from the calling context, and you can likely leave it as-is.