Split a vector into a table of place values

Generate a table of places (by power of 10, with column names such as "10^2" and "10^-2") from a vector of numeric values.

Usage

splitByPlace(
  x,
  max_power_10 = 2,
  min_power_10 = -2,
  x_arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)

Arguments

x

numeric A vector of values to evaluate.

max_power_10

integer The highest 10s place to check, as a power of 10. This will set the first column of the returned results.

min_power_10

integer The highest 10s place to check, as a power of 10. Must be lower than max_power_10. This will set the first column of the returned results.

x_arg

character The name of x to use in the output table and error messages. By default this is automatically deteermined from the calling context, and you can likely leave it as-is.

call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.

Value

A tibble::tibble() with a column for the original vector (named using x_arg) and columns for each power of 10 between max_power_10 and min_power_10.

Examples

random_data <- runif(10, 0, 1000)
splitByPlace(random_data)
#> # A tibble: 10 × 6
#>    random_data `10^2` `10^1` `10^0` `10^-1` `10^-2`
#>          <dbl>  <int>  <int>  <int>   <int>   <int>
#>  1      80.750     NA      8      0       7       5
#>  2     834.333      8      3      4       3       3
#>  3     600.761      6      0      0       7       6
#>  4     157.208      1      5      7       2       0
#>  5       7.399     NA     NA      7       3       9
#>  6     466.393      4      6      6       3       9
#>  7     497.777      4      9      7       7       7
#>  8     289.767      2      8      9       7       6
#>  9     732.882      7      3      2       8       8
#> 10     772.522      7      7      2       5       2
splitByPlace(random_data, 2, -5)
#> # A tibble: 10 × 9
#>    random_data `10^2` `10^1` `10^0` `10^-1` `10^-2` `10^-3` `10^-4` `10^-5`
#>          <dbl>  <int>  <int>  <int>   <int>   <int>   <int>   <int>   <int>
#>  1   80.750138     NA      8      0       7       5       0       1       3
#>  2  834.333037      8      3      4       3       3       3       0       3
#>  3  600.760886      6      0      0       7       6       0       8       8
#>  4  157.208442      1      5      7       2       0       8       4       4
#>  5    7.399441     NA     NA      7       3       9       9       4       4
#>  6  466.393497      4      6      6       3       9       3       4       9
#>  7  497.777389      4      9      7       7       7       7       3       8
#>  8  289.767245      2      8      9       7       6       7       2       4
#>  9  732.881987      7      3      2       8       8       1       9       8
#> 10  772.521511      7      7      2       5       2       1       5       1
splitByPlace(random_data, 2, -5, "measurements")
#> # A tibble: 10 × 9
#>    measurements `10^2` `10^1` `10^0` `10^-1` `10^-2` `10^-3` `10^-4` `10^-5`
#>           <dbl>  <int>  <int>  <int>   <int>   <int>   <int>   <int>   <int>
#>  1    80.750138     NA      8      0       7       5       0       1       3
#>  2   834.333037      8      3      4       3       3       3       0       3
#>  3   600.760886      6      0      0       7       6       0       8       8
#>  4   157.208442      1      5      7       2       0       8       4       4
#>  5     7.399441     NA     NA      7       3       9       9       4       4
#>  6   466.393497      4      6      6       3       9       3       4       9
#>  7   497.777389      4      9      7       7       7       7       3       8
#>  8   289.767245      2      8      9       7       6       7       2       4
#>  9   732.881987      7      3      2       8       8       1       9       8
#> 10   772.521511      7      7      2       5       2       1       5       1